# [Builtins] [Evaluation] Drop lookups