--- a/plutus-core/cost-model/budgeting-bench/Common.hs
+++ b/plutus-core/cost-model/budgeting-bench/Common.hs
@@ -23,7 +23,6 @@ import UntypedPlutusCore.Evaluation.Machine.Cek
 import Control.DeepSeq (NFData, force)
 import Criterion.Main
 import Data.ByteString qualified as BS
-import Data.Ix (Ix)
 import Data.Typeable (Typeable)
 
 type PlainTerm uni fun = UPLC.Term Name uni fun ()
@@ -74,7 +73,7 @@ copyData =
 ---------------- Creating benchmarks ----------------
 
 benchWith
-    :: (Ix fun, Pretty fun, Typeable fun)
+    :: (Pretty fun, Typeable fun)
     => MachineParameters CekMachineCosts CekValue DefaultUni fun ()
     -> String
     -> PlainTerm DefaultUni fun