--- a/plutus-core/plutus-core/src/PlutusCore/Builtin/Meaning.hs
+++ b/plutus-core/plutus-core/src/PlutusCore/Builtin/Meaning.hs
@@ -31,11 +31,12 @@ import PlutusCore.Evaluation.Machine.ExBudget
 import PlutusCore.Evaluation.Machine.ExMemory
 import PlutusCore.Name
 
+import Control.DeepSeq
 import Data.Array
 import Data.Kind qualified as GHC
 import Data.Proxy
 import Data.Some.GADT
-import GHC.Exts (inline, oneShot)
+import GHC.Exts (inline, lazy, oneShot)
 import GHC.TypeLits
 
 -- | Turn a list of Haskell types @args@ into a functional type ending in @res@.
@@ -126,7 +127,7 @@ denotation is.
 'KnownPolytype' and 'KnownMonotype' are responsible for deriving polymorphic and monomorphic types,
 respectively.
 
-'KnownPolytype' turns every bound variable into a 'TypeSchemeAll'/'RuntimeSchemeAll'. We extract
+'KnownPolytype' turns every bound variable into a 'TypeSchemeAll'/'BuiltinExpectForce'. We extract
 variables from the type of the Haskell denotation using the 'ToBinds' associated type
 family. Variables are collected in the order that they appear in (i.e. just like in Haskell). For
 example, processing a type like
@@ -141,7 +142,7 @@ with 'ToBinds' results in the following list of bindings:
     '[ 'Some ('TyNameRep "b" 1), 'Some ('TyNameRep "a" 0) ]
 
 'KnownMonotype' turns every argument that the Haskell denotation of a builtin receives into a
-'TypeSchemeArrow'/'RuntimeSchemeArrow'. We extract the arguments from the type of the Haskell
+'TypeSchemeArrow'/'BuiltinExpectArgument'. We extract the arguments from the type of the Haskell
 denotation using the 'GetArgs' type family.
 
 Higher-kinded type variables are fully supported.
@@ -261,7 +262,7 @@ instance
         -- Ironically computing the unlifted value strictly is the best way of doing deferred
         -- unlifting. This means that while the resulting 'ReadKnownM' is only handled upon full
         -- saturation and any evaluation failure is only registered when the whole builtin
-        -- application is evaluated.
+        -- application is evaluated, a Haskell exception will occur immediately.
         -- It shouldn't matter though, because a builtin is not supposed to throw an
         -- exception at any stage, that would be a bug regardless.
         toMonoF @val @args @res $! do
@@ -343,11 +344,17 @@ instance
         , ElaborateFromTo 0 j val a, KnownPolytype binds val args res
         ) => MakeBuiltinMeaning a val where
     makeBuiltinMeaning f toExF =
-        BuiltinMeaning (knownPolytype @binds @val @args @res) f $
-            -- See Note [Optimizations of runCostingFun*] for why we use strict @case@.
-            \cost ->
-                case toExF cost of
-                    !exF -> toPolyF @binds @val @args @res $ pure (f, exF)
+        BuiltinMeaning (knownPolytype @binds @val @args @res) f $ \cost ->
+            -- In order to make the 'BuiltinRuntime' of a builtin cacheable we need to tell GHC to
+            -- create a thunk for it, which we achieve by applying 'lazy' to the 'BuiltinRuntime'
+            -- here.
+            --
+            -- Those thunks however require a lot of care to be properly shared rather than
+            -- recreated every time a builtin application is evaluated, see 'toBuiltinsRuntime' for
+            -- how we sort it out.
+            lazy $ case toExF cost of
+                -- See Note [Optimizations of runCostingFun*] for why we use strict @case@.
+                !exF -> toPolyF @binds @val @args @res $ pure (f, exF)
     {-# INLINE makeBuiltinMeaning #-}
 
 -- | Convert a 'BuiltinMeaning' to a 'BuiltinRuntime' given a cost model.
@@ -362,7 +369,22 @@ toBuiltinsRuntime
     :: (cost ~ CostingPart uni fun, ToBuiltinMeaning uni fun, HasMeaningIn uni val)
     => BuiltinVersion fun -> cost -> BuiltinsRuntime fun val
 toBuiltinsRuntime ver cost =
-    let arr = tabulateArray $ toBuiltinRuntime cost . inline toBuiltinMeaning ver
-    in -- Force array elements to WHNF
-        foldr seq (BuiltinsRuntime arr) arr
+    let runtime = BuiltinsRuntime $ toBuiltinRuntime cost . inline toBuiltinMeaning ver
+        {-# INLINE runtime #-}
+       -- Force each 'BuiltinRuntime' to WHNF. Inlining 'force' manually, since it doesn't have an
+       -- @INLINE@ pragma. This allows GHC to get to the 'NFData' instance for 'BuiltinsRuntime',
+       -- which forces all the freshly created 'BuiltinRuntime' thunks. Which is important, because
+       -- the thunks are behind a lambda binding the @cost@ variable and GHC would supply the @cost@
+       -- value (the one that is in the current scope) at runtime, if we didn't tell it that the
+       -- thunks need to be forced early. Which would be detrimental to performance, since it would
+       -- mean that the thunks would be created at runtime over and over again, each time we go
+       -- under the lambda binding the @cost@ variable, i.e. each time the 'BuiltinRuntime' is
+       -- retrieved from the environment. The 'deepseq' nagging causes GHC to supply the @cost@
+       -- value at compile time, thus allocating the thunks within this entire function allowing
+       -- them to be reused each time the 'BuiltinRuntime' is looked up (after the initial phase
+       -- forcing all of them at once).
+       --
+       -- Note that despite @runtime@ being used twice, we don't get all the multiple thousands of
+       -- Core duplicated, because the 'BuiltinRuntime' thunks are shared in the two @runtime@s.
+    in runtime `deepseq` runtime
 {-# INLINE toBuiltinsRuntime #-}