--- a/plutus-core/plutus-core/src/PlutusCore/Evaluation/Machine/MachineParameters/Default.hs
+++ b/plutus-core/plutus-core/src/PlutusCore/Evaluation/Machine/MachineParameters/Default.hs
@@ -45,6 +45,6 @@ mkMachineParametersFor :: (MonadError CostModelApplyError m)
 mkMachineParametersFor ver newCMP =
     inline mkMachineParameters ver <$>
         applyCostModelParams defaultCekCostModel newCMP
--- {-# INLINE mkMachineParametersFor #-} was removed because [benchmarking
--- results](https://github.com/input-output-hk/plutus/pull/4879#issuecomment-1301052379) show that
--- the pragma isn't helping anymore.
+-- Not marking this function with @INLINE@, since at this point everything we wanted to be inlined
+-- is inlined and there's zero reason to duplicate thousands and thousands of lines of Core down
+-- the line.