--- a/plutus-core/untyped-plutus-core/src/UntypedPlutusCore/Evaluation/Machine/Cek.hs
+++ b/plutus-core/untyped-plutus-core/src/UntypedPlutusCore/Evaluation/Machine/Cek.hs
@@ -72,7 +72,6 @@ import PlutusCore.Quote
 import Control.Monad.Except
 import Control.Monad.State
 import Data.Bifunctor
-import Data.Ix (Ix)
 import Data.Text (Text)
 import Universe
 
@@ -92,7 +91,7 @@ A wrapper around the internal runCek to debruijn input and undebruijn output.
 *THIS FUNCTION IS PARTIAL if the input term contains free variables*
 -}
 runCek
-    :: (Ix fun, PrettyUni uni fun)
+    :: PrettyUni uni fun
     => MachineParameters CekMachineCosts CekValue uni fun ann
     -> ExBudgetMode cost uni fun
     -> EmitterMode uni fun
@@ -123,7 +122,7 @@ runCek params mode emitMode term =
 -- | Evaluate a term using the CEK machine with logging disabled and keep track of costing.
 -- *THIS FUNCTION IS PARTIAL if the input term contains free variables*
 runCekNoEmit
-    :: (Ix fun, PrettyUni uni fun)
+    :: PrettyUni uni fun
     => MachineParameters CekMachineCosts CekValue uni fun ann
     -> ExBudgetMode cost uni fun
     -> Term Name uni fun ann
@@ -139,7 +138,7 @@ May throw a 'CekMachineException'.
 unsafeRunCekNoEmit
     :: ( Pretty (SomeTypeIn uni), Typeable uni
        , Closed uni, uni `Everywhere` PrettyConst
-       , Ix fun, Pretty fun, Typeable fun
+       , Pretty fun, Typeable fun
        )
     => MachineParameters CekMachineCosts CekValue uni fun ann
     -> ExBudgetMode cost uni fun
@@ -152,7 +151,7 @@ unsafeRunCekNoEmit params mode =
 -- | Evaluate a term using the CEK machine with logging enabled.
 -- *THIS FUNCTION IS PARTIAL if the input term contains free variables*
 evaluateCek
-    :: (Ix fun, PrettyUni uni fun)
+    :: PrettyUni uni fun
     => EmitterMode uni fun
     -> MachineParameters CekMachineCosts CekValue uni fun ann
     -> Term Name uni fun ann
@@ -164,7 +163,7 @@ evaluateCek emitMode params =
 -- | Evaluate a term using the CEK machine with logging disabled.
 -- *THIS FUNCTION IS PARTIAL if the input term contains free variables*
 evaluateCekNoEmit
-    :: (Ix fun, PrettyUni uni fun)
+    :: PrettyUni uni fun
     => MachineParameters CekMachineCosts CekValue uni fun ann
     -> Term Name uni fun ann
     -> Either (CekEvaluationException Name uni fun) (Term Name uni fun ())
@@ -175,7 +174,7 @@ evaluateCekNoEmit params = fst . runCekNoEmit params restrictingEnormous
 unsafeEvaluateCek
     :: ( Pretty (SomeTypeIn uni), Typeable uni
        , Closed uni, uni `Everywhere` PrettyConst
-       , Ix fun, Pretty fun, Typeable fun
+       , Pretty fun, Typeable fun
        )
     => EmitterMode uni fun
     -> MachineParameters CekMachineCosts CekValue uni fun ann
@@ -190,7 +189,7 @@ unsafeEvaluateCek emitTime params =
 unsafeEvaluateCekNoEmit
     :: ( Pretty (SomeTypeIn uni), Typeable uni
        , Closed uni, uni `Everywhere` PrettyConst
-       , Ix fun, Pretty fun, Typeable fun
+       , Pretty fun, Typeable fun
        )
     => MachineParameters CekMachineCosts CekValue uni fun ann
     -> Term Name uni fun ann
@@ -201,7 +200,7 @@ unsafeEvaluateCekNoEmit params = unsafeExtractEvaluationResult . evaluateCekNoEm
 -- *THIS FUNCTION IS PARTIAL if the input term contains free variables*
 readKnownCek
     :: ( ReadKnown (Term Name uni fun ()) a
-       , Ix fun, PrettyUni uni fun
+       , PrettyUni uni fun
        )
     => MachineParameters CekMachineCosts CekValue uni fun ann
     -> Term Name uni fun ann