@effectfully 2022-10-20
Running the benchmarks manually:

<details>
<summary>Results table</summary>

| Script                                   |  Old       |  New       |   Change   |
| :------                                  |  :------:  |  :------:  |  :------:  |
| auction_1-1                              |  178.0 μs  |  149.9 μs  |   -16,3%   |
| auction_1-2                              |  686.6 μs  |  664.9 μs  |    -3,2%   |
| auction_1-3                              |  693.2 μs  |  653.8 μs  |    -5,8%   |
| auction_1-4                              |  233.2 μs  |  199.0 μs  |   -14,6%   |
| auction_2-1                              |  178.9 μs  |  154.0 μs  |   -13,5%   |
| auction_2-2                              |  691.8 μs  |  660.6 μs  |    -4,5%   |
| auction_2-3                              |  893.6 μs  |  850.1 μs  |    -4,8%   |
| auction_2-4                              |  690.5 μs  |  652.3 μs  |    -5,5%   |
| auction_2-5                              |  229.6 μs  |  197.6 μs  |   -14,0%   |
| crowdfunding-success-1                   |  206.9 μs  |  181.7 μs  |   -12,1%   |
| crowdfunding-success-2                   |  209.7 μs  |  182.5 μs  |   -12,9%   |
| crowdfunding-success-3                   |  207.2 μs  |  180.7 μs  |   -13,0%   |
| currency-1                               |  251.6 μs  |  234.9 μs  |    -6,8%   |
| escrow-redeem_1-1                        |  368.0 μs  |  332.9 μs  |    -9,8%   |
| escrow-redeem_1-2                        |  366.3 μs  |  332.9 μs  |    -9,3%   |
| escrow-redeem_2-1                        |  440.2 μs  |  396.3 μs  |   -10,0%   |
| escrow-redeem_2-2                        |  436.1 μs  |  397.0 μs  |    -8,9%   |
| escrow-redeem_2-3                        |  435.2 μs  |  395.8 μs  |    -9,2%   |
| escrow-refund-1                          |  156.8 μs  |  137.2 μs  |   -12,2%   |
| future-increase-margin-1                 |  252.5 μs  |  236.1 μs  |    -6,3%   |
| future-increase-margin-2                 |  578.6 μs  |  540.3 μs  |    -6,6%   |
| future-increase-margin-3                 |  576.1 μs  |  538.1 μs  |    -6,6%   |
| future-increase-margin-4                 |  542.8 μs  |  507.9 μs  |    -6,5%   |
| future-increase-margin-5                 |  861.9 μs  |  830.4 μs  |    -3,6%   |
| future-pay-out-1                         |  254.0 μs  |  234.9 μs  |    -7,9%   |
| future-pay-out-2                         |  575.9 μs  |  540.0 μs  |    -6,1%   |
| future-pay-out-3                         |  576.9 μs  |  543.7 μs  |    -5,7%   |
| future-pay-out-4                         |  864.1 μs  |  832.8 μs  |    -3,7%   |
| future-settle-early-1                    |  253.6 μs  |  236.2 μs  |    -6,7%   |
| future-settle-early-2                    |  579.8 μs  |  540.7 μs  |    -6,7%   |
| future-settle-early-3                    |  579.7 μs  |  540.0 μs  |    -6,7%   |
| future-settle-early-4                    |  656.2 μs  |  628.8 μs  |    -4,3%   |
| game-sm-success_1-1                      |  416.8 μs  |  379.5 μs  |    -8,9%   |
| game-sm-success_1-2                      |  198.9 μs  |  167.0 μs  |   -15,7%   |
| game-sm-success_1-3                      |  682.5 μs  |  653.0 μs  |    -4,3%   |
| game-sm-success_1-4                      |  226.0 μs  |  195.3 μs  |   -13,7%   |
| game-sm-success_2-1                      |  416.4 μs  |  377.8 μs  |    -9,4%   |
| game-sm-success_2-2                      |  198.3 μs  |  168.4 μs  |   -15,2%   |
| game-sm-success_2-3                      |  686.8 μs  |  653.9 μs  |    -4,8%   |
| game-sm-success_2-4                      |  231.7 μs  |  194.8 μs  |   -16,0%   |
| game-sm-success_2-5                      |  685.1 μs  |  654.4 μs  |    -4,5%   |
| game-sm-success_2-6                      |  232.2 μs  |  194.2 μs  |   -16,4%   |
| multisig-sm-1                            |  424.8 μs  |  392.7 μs  |    -7,5%   |
| multisig-sm-2                            |  415.0 μs  |  387.1 μs  |    -6,7%   |
| multisig-sm-3                            |  423.9 μs  |  359.3 μs  |   -15,1%   |
| multisig-sm-4                            |  428.5 μs  |  394.6 μs  |    -7,9%   |
| multisig-sm-5                            |  608.4 μs  |  584.2 μs  |    -3,9%   |
| multisig-sm-6                            |  423.0 μs  |  393.5 μs  |    -7,1%   |
| multisig-sm-7                            |  414.9 μs  |  383.9 μs  |    -7,5%   |
| multisig-sm-8                            |  424.9 μs  |  359.7 μs  |   -15,3%   |
| multisig-sm-9                            |  430.3 μs  |  397.9 μs  |    -7,7%   |
| multisig-sm-10                           |  613.2 μs  |  583.0 μs  |    -4,9%   |
| ping-pong-1                              |  348.6 μs  |  321.0 μs  |    -7,8%   |
| ping-pong-2                              |  348.0 μs  |  320.1 μs  |    -8,0%   |
| ping-pong_2-1                            |  205.8 μs  |  184.9 μs  |   -10,2%   |
| prism-1                                  |  165.8 μs  |  140.8 μs  |   -15,2%   |
| prism-2                                  |  448.1 μs  |  414.7 μs  |    -7,6%   |
| prism-3                                  |  379.8 μs  |  345.4 μs  |    -9,0%   |
| pubkey-1                                 |  137.8 μs  |  119.3 μs  |   -13,1%   |
| stablecoin_1-1                           |  949.2 μs  |  901.7 μs  |    -5,1%   |
| stablecoin_1-2                           |  189.9 μs  |  162.7 μs  |   -14,3%   |
| stablecoin_1-3                           |  1.082 ms  |  1.023 ms  |     0,0%   |
| stablecoin_1-4                           |  200.3 μs  |  174.9 μs  |   -13,0%   |
| stablecoin_1-5                           |  1.348 ms  |  1.268 ms  |     0,0%   |
| stablecoin_1-6                           |  250.4 μs  |  218.4 μs  |   -12,8%   |
| stablecoin_2-1                           |  947.1 μs  |  901.5 μs  |    -4,9%   |
| stablecoin_2-2                           |  190.1 μs  |  163.5 μs  |   -14,2%   |
| stablecoin_2-3                           |  1.084 ms  |  1.022 ms  |     0,0%   |
| stablecoin_2-4                           |  202.0 μs  |  173.8 μs  |   -14,4%   |
| token-account-1                          |  190.6 μs  |  171.6 μs  |   -10,0%   |
| token-account-2                          |  337.7 μs  |  315.2 μs  |    -6,5%   |
| uniswap-1                                |  432.7 μs  |  415.2 μs  |    -3,9%   |
| uniswap-2                                |  226.2 μs  |  204.8 μs  |    -9,7%   |
| uniswap-3                                |  1.707 ms  |  1.615 ms  |     0,0%   |
| uniswap-4                                |  336.5 μs  |  291.3 μs  |   -13,4%   |
| uniswap-5                                |  1.180 ms  |  1.096 ms  |     0,0%   |
| uniswap-6                                |  323.5 μs  |  279.9 μs  |   -13,6%   |
| vesting-1                                |  364.7 μs  |  343.3 μs  |    -5,8%   |
</details>