@effectfully 2022-10-20
^ -8.57% on average. I was unable to fall asleep and that gave me enough time to finally realize how to remove those damn lookups.